+++ spec/features/terms_of_service_spec.rb
@@ -3,34 +3,46 @@
 require 'rails_helper'
 
 RSpec.feature 'Terms of service' do
-  context 'as a student' do
-    let(:user) { create(:user, tos_accepted: nil) }
+  describe 'acceptance' do
+    context 'as a student' do
+      let(:user) { create(:user, tos_accepted: nil) }
 
-    it 'the form can be accepted' do
-      log_in user
-      click_on 'Accept Terms of Service'
-      expect(page).to have_css('.flash-success', text: /Welcome to Vesta/)
+      it 'the form can be accepted' do
+        log_in user
+        click_on 'Accept Terms of Service'
+        expect(page).to have_css('.flash-success', text: /Welcome to Vesta/)
+      end
+      it 'pages redirect to the tos if it is not accepted' do
+        log_in user
+        visit user_path(user)
+        expect(page).to have_content('Terms of Service')
+      end
+      it 'will not redirect if the terms of service is accepted' do
+        user.update!(tos_accepted: Time.current)
+        log_in user
+        visit user_path(user)
+        msg = 'You must accept the Terms of Service to proceed.'
+        expect(page).not_to have_content(msg)
+      end
     end
-    it 'pages redirect to the tos if it is not accepted' do
-      log_in user
-      visit user_path(user)
-      expect(page).to have_content('Terms of Service')
-    end
-    it 'will not redirect if the terms of service is accepted' do
-      user.update!(tos_accepted: Time.current)
-      log_in user
-      visit user_path(user)
-      msg = 'You must accept the Terms of Service to proceed.'
-      expect(page).not_to have_content(msg)
+    context 'as an admin' do
+      let(:user) { create(:admin) }
+
+      it 'the acceptance link does not appear' do
+        log_in user
+        visit terms_of_service_path
+        expect(page).not_to have_link('Accept Terms of Service')
+      end
     end
   end
-  context 'as an admin' do
-    let(:user) { create(:admin) }
 
-    it 'the acceptance link does not appear' do
-      log_in user
-      visit terms_of_service_path
-      expect(page).not_to have_link('Accept Terms of Service')
+  describe 'resetting' do
+    it 'allows superusers to reset all terms of service acceptances' do
+      log_in create(:admin, role: 'superuser')
+      create(:user, role: 'student')
+      click_on 'Settings'
+      click_on 'Reset terms of service acceptance for all non-admins'
+      expect(page).to have_css('.flash-success', text: /has been reset/)
     end
   end
 end