+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 * Add the ability to assign lottery numbers by size ([#848](https://gitlab.com/yale-sdmp/vesta/issues/848)).
 * Add admins to all bulk user emails ([#887](https://gitlab.com/yale-sdmp/vesta/issues/887)).
+* Add ability to reset terms of service acceptance ([#904](https://gitlab.com/yale-sdmp/vesta/issues/904)).
 
 ### Fixed
 * Scope the results view and export to only return active students in the current college ([#885](https://gitlab.com/yale-sdmp/vesta/issues/885)).